Renamed files of renamed operations.  Added alias
Added alias for fetchDeployments as it is currently used by the CLI
